+++ paper.bib
@@ -7,7 +7,8 @@
   number={7857},
   pages={74--82},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41586-021-03302-y}
 }
 
 @article{Favier2014,
@@ -18,7 +19,8 @@
   number={2},
   pages={117--121},
   year={2014},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nclimate2094}
 }
 
 
@@ -30,7 +32,8 @@
   number={7857},
   pages={83--89},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41586-021-03427-0}
 }
 
 @article{DeRydt2021,
@@ -41,7 +44,8 @@
   number={1},
   pages={113--132},
   year={2021},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/tc-15-113-2021}
 }
 
 @software{Shapero,
@@ -68,7 +72,8 @@
   number={24},
   pages={eabg3080},
   year={2021},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/sciadv.abg3080}
 }
 
 @article{Schoof2007,
@@ -78,7 +83,8 @@
   volume={112},
   number={F3},
   year={2007},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2006JF000664}
 }
 
 @article{Arthern2015,
@@ -89,7 +95,8 @@
   number={7},
   pages={1171--1188},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2014JF003239}
 }
 
 @article{Arthern2017,
@@ -100,7 +107,8 @@
   number={5},
   pages={2352--2359},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2017GL072514}
 }
 
 @article{Goldberg2011,
@@ -111,7 +119,8 @@
   number={201},
   pages={157--170},
   year={2011},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.3189/002214311795306763}
 }
 
 @article{AsayDavis2017,
@@ -122,7 +131,8 @@
   number={4},
   pages={316--329},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s40641-017-0071-0}
 }
 
 @article{Marshall1997,
@@ -133,7 +143,8 @@
   number={C3},
   pages={5733--5752},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/96JC02776}
 }
 
 @article{Vasilyev2005,
@@ -144,7 +155,8 @@
   number={2},
   pages={412--431},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2004.12.013}
 }
 
 @article{Cornford2020,
@@ -155,7 +167,8 @@
   number={7},
   pages={2283--2301},
   year={2020},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/tc-14-2283-2020}
 }
 
 @article{Larour2012,
@@ -165,7 +178,8 @@
   volume={117},
   number={F1},
   year={2012},
-  publisher={Wiley Online Library}
[…]
