### 2023-05-09 3faa347
fix grid size definition in WAIS example
