+++ paper.md
@@ -18,7 +18,10 @@ affiliations:
 date: 30 Sept 2022
 bibliography: paper.bib
 ---
+# Summary 
+Ice sheet models are used to improve our understanding of the past, present, and future evolution of ice sheets. To do so, they solve the equations describing the flow of ice when forced by other climate elements, particularly the atmosphere and oceans. We present `WAVI.jl`, an ice sheet model written in Julia. `WAVI.jl` is designed to lower the bar to ice sheet modelling, whilst including sufficient detail to be used for addressing cutting-edge research questions. 
 
+# Statement of Need
 Ice sheet models allow us to simulate the behaviour and evolution of ice sheets, which are large masses of glacial land and marine ice. There are two main uses of ice sheet models: firstly, prognostic use, which involves making predictions about the future of ice sheets. Prognostic predictions often relate to sea level rise contributions: the world's two largest ice sheets, located in Antarctica and Greenland, hold enough ice to raise sea levels by approximately 58 and 7 meters, respectively [@Bamber2018]; prognostic modelling of these ice sheets enables us to make predictions on (for example) how much of this ice will be lost [@Edwards2021], to investigate the possibility of runaway ice loss [@DeConto2021], and analyze whether or not such instabilities have already been initiated [@Favier2014]. Secondly, ice sheet models are also used diagnostically, which involves using a model to investigate processes controlling the behaviour of an ice sheet, such as how loss of ice shelves -- the floating extensions of ice sheets -- influences ice flow speed [@Joughin2021], how different bed conditions affect ice sliding [@DeRydt2021], and probing the conditions under which so-called 'tipping-points' might be passed [@Schoof2007].
 
 On the long [O(1000s km)] lengthscales that are relevant to ice sheets, ice behaves approximately as a highly viscous fluid with a shear-thinning rheology (meaning that as the ice deforms, it becomes thinner and flows more easily). `WAVI.jl` (Wavelet-based Adaptive-grid Vertically-integrated Ice-sheet-model) is a Julia package for the numerical solution of an accurate approximation to the Stokes equations, which describe conservation of mass and momentum in such a fluid. This approximation, which is appropriate for fluid flows with a high aspect ratio (as is the case for the vast majority of ice sheets), treats longitudinal and lateral stresses as depth independent, but accounts for vertical velocity gradients in the nonlinear viscosity and in the treatment of basal stress [@Goldberg2011].
