### 2023-05-15 #5452 @kthyng
@alextbradley Could you take a look at the reviewer list or this website (https://reviewers.joss.theoj.org/) and suggest 5 reviewers? Please don't ping them with "@", just list their handles.
