### 2023-06-22 #5584 @editorialbot
[REVIEW]: WAVI.jl: Ice Sheet Modelling in Julia
**Submitting author:** @alextbradley (<a href="http://orcid.org/0000-0001-8381-5317">Alexander Bradley</a>)
**Repository:** https://github.com/RJArthern/WAVI.jl
**Branch with paper.md** (empty if default branch): build-docs
**Version:** v0.0.2
