### 2023-06-28 #5584 @PennyHow
Hi @alextbradley et al. I've just had a look through your repo and the paper. Nice work! Overall I think this is a really valuable alternative to ice sheet models currently available. I don't think I have seen any openly available ice sheet models in Julia, so I think this is a first!

I have some queries about running the model and its documentation. From initially setting up Julia and downloading `WAVI.jl` through your documented set-up, I have some first comments.

**Concerning the software**

I am having problems running the MISMIP+ experiment outlined in your README.
1. Should this line have some bracketed component?
```
bed = WAVI.mismip_plus_bed
```

2. And secondly, this line is throwing up this method error.
```
julia> model = Model(grid = grid,bed_elevation = bed, params = params)
ERROR: MethodError: no method matching WAVI.HGrid(::Int64, ::Int64, ::Matrix{Bool}, ::Int64, ::LinearAlgebra.Diagonal{Float64, Vector{Float64}}, ::Matrix{Float64}, ::SparseArrays.SparseMatrixCSC{Float64, Int64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Base.RefValue{LinearAlgebra.Diagonal{Float64, Vector{Float64}}}, ::Base.RefValue{LinearAlgebra.Diagonal{Float64, Vector{Float64}}})

Closest candidates are:
  WAVI.HGrid(::N, ::N, ::Matrix{Bool}, ::N, ::LinearAlgebra.Diagonal{T, Vector{T}}, ::SparseArrays.SparseMatrixCSC{T, N}, ::SparseArrays.SparseMatrixCSC{T, N}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Base.RefValue{LinearAlgebra.Diagonal{T, Vector{T}}}, ::Base.RefValue{LinearAlgebra.Diagonal{T, Vector{T}}}) where {T<:Real, N<:Integer}
   @ WAVI ~/.julia/packages/WAVI/osoCg/src/Fields/HGrid.jl:2

Stacktrace:e2 [1] WAVI.HGrid(; nxh::Int64, nyh::Int64, mask::Matrix{Bool}, b::Matrix{Float64}, h::Matrix{Float64}, ηav::Matrix{Float64})
   @ WAVI ~/.julia/packages/WAVI/osoCg/src/Fields/HGrid.jl:144
 [2] setup_fields(grid::Grid{Float64, Int64}, initial_conditions::InitialConditions{Float64}, solver_params::SolverParams{Float64, Int64}, params::Params{Float64, Matrix{Float64}, Matrix{Float64}, Matrix{Float64}}, bed_array::Matrix{Float64})
   @ WAVI ~/.julia/packages/WAVI/osoCg/src/Fields/Fields.jl:41
 [3] Model(; grid::Grid{Float64, Int64}, bed_elevation::Function, params::Params{Float64, Float64, Float64, Float64}, solver_params::SolverParams{Float64, Int64}, initial_conditions::InitialConditions{Float64}, melt_rate::UniformMeltRate{Float64})
   @ WAVI ~/.julia/packages/WAVI/osoCg/src/Models/Model.jl:83
 [4] top-level scope
   @ REPL[13]:1
```

I checked the other examples on your really nicely documentation (https://rjarthern.github.io/WAVI.jl/). And I came across the same problem with this line in each of the examples. I'm not as familiar with Julia, so is this something I am doing wrong, something else I need in my set-up, or something that needs to be modified in your examples? Just FYI, I'm running Julia 1.9.0 on a Ubuntu 22.04 LTS machine.

**Concerning the documentation**

I had to search hard for your documentation at https://rjarthern.github.io/WAVI.jl/ as I don't think it is included in your repo's README. Could you add this as a link to the repo, or a badge (if this is possible with your `Documenter.jl `set-up)? It has some great theoretical summaries, examples and walkthroughs.

With regard to the documentation, I have a request which I am happy to discuss with you the best way forward - Most packages use some form of autodocs to automatically generate documentation from the code (e.g. function arguments, outputs, and brief description). This would be highly beneficial to users to understand the capabilities and inputs to functions, troubleshoot problems, and adapt functions where needed. Could this be a possibility for you to include in your documentation?

**Concerning the software paper**

Great, easy-to-follow paper with a fantastic overview of current ice sheet modelling theory and available models. Just a couple of comments from me on this...

- L8: I would say "lower the bar" is quite a colloqial term. Please change this to something more like "`WAVI.jl` is designed to make ice sheet modelling more accessible to beginners and low-level users..."
- L25: What does the bracketed part of "On the long [O(1000s km)]..." mean"
- L25: "lengthscales" >> "length scales"
