### 2023-06-30 #5584 @observingClouds
Hi @PennyHow,
Hi @daniel-cheng,
I see you are busy working through the checklist and have already made the first comments. It looks like we are all on track. Please be reminded to also make use of the issue functionality of GitHub and file issues on WAV.jl's repository. Discussions can quickly go out of hand on this thread here if there is a lot of back and forth.

@alextbradley, please feel free to already address some of the posted comments. If they require a longer response, I recommend opening issues on your end and tag the specific reviewer over there to continue the discussion.

Cheers!
