### 2023-07-05 #5584 @JordiBolibar
# Review for WAVI.jl

Hi @alextbradley et al., I have finished reading your manuscript, repository and docs. Overall, great work and extremely happy to witness the first ice sheet model in Julia! I think the goal of the model is well presented, the API looks clean, and the docs are quite exhaustive. Nonetheless, I've come across some technical issues when trying to use the model, and I also have some feedback to potentially improve the docs.

## Installation

During installation, following your guidelines, I encountered the following warning:

```julia
1 dependency had warnings during precompilation:
┌ WAVI [881191c6-26ca-4787-919d-8f98f1c41532]
│  WARNING: Method definition spdiagm(Integer, Integer, (Pair{var"#s87", var"#s86"} where var"#s86"<:(AbstractArray{T, 1} where T) where var"#s87"<:Integer)...) in module SparseArrays at /cache/build/default-amdci4-6/julialang/julia-release-1-dot-9/usr/share/julia/stdlib/v1.9/SparseArrays/src/sparsematrix.jl:3986 overwritten in module WAVI at /home/jovyan/.julia/packages/WAVI/osoCg/src/WAVI.jl:68.
│	** incremental compilation may be fatally broken for this module **
└
```

There seems to be an issue with the package SparseArrays.jl, which as I will discuss in the next section, also hinders the use of your package.

## Running the code

I have tried running some of your examples (both the one in the README and the simplest one in the docs), but the code seems to crash when initializing the grid:

```julia
grid = Grid(nx = 300, ny = 2, dx = 12000.0, dy = 12000.0)
```
which produces:

```julia
ERROR: MethodError: no method matching WAVI.HGrid(::Int64, ::Int64, ::Matrix{Bool}, ::Int64, ::LinearAlgebra.Diagonal{Float64, Vector{Float64}}, ::Matrix{Float64}, ::SparseArrays.SparseMatrixCSC{Float64, Int64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Matrix{Float64}, ::Base.RefValue{LinearAlgebra.Diagonal{Float64, Vector{Float64}}}, ::Base.RefValue{LinearAlgebra.Diagonal{Float64, Vector{Float64}}})

Closest candidates are:
  WAVI.HGrid(::N, ::N, ::Matrix{Bool}, ::N, ::LinearAlgebra.Diagonal{T, Vector{T}}, ::SparseArrays.SparseMatrixCSC{T, N}, ::SparseArrays.SparseMatrixCSC{T, N}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Matrix{T}, ::Base.RefValue{LinearAlgebra.Diagonal{T, Vector{T}}}, ::Base.RefValue{LinearAlgebra.Diagonal{T, Vector{T}}}) where {T<:Real, N<:Integer}
   @ WAVI ~/.julia/packages/WAVI/osoCg/src/Fields/HGrid.jl:2
```
I would say this is by far the most important issue to address in the review. Probably it's just a minor technical issue  with the package, but it would be important to have everything running smoothly before widely sharing the library with the community.

## Authors

I'm not familiar with the format of the authors in JOSS, but to me in the manuscript there seems to be 3 authors, and in the repository and the code there seems to be mainly two authors. Is this normal? Probably I'm misunderstanding something here, so apologies if I got this wrong.

## Documentation

### Access

I see this was also mentioned by @PennyHow, but I found it really hard to find the docs. Actually, I randomly came across them some weeks ago, but when I tried to look for the link in the repository during the review I just couldn't find it. I think a link and a mention of the docs should be clearly stated in the README (perhaps with a docs badge), and/or in the repository description.

### API

The WAVI Setup section name is a little bit confusing to me. I feel it would be clearer to call it "API" or "WAVI data structures". Setup sounds more like configuration or installation to me.

The image showing the data structures hierarchy is not displaying in the docs. It would be nice to fix this, since I'm sure it's really useful to users.

The simulations section has some formatting issues, where some code is displayed as a title.

### Physics

I find the term "melt rates" a bit vague. After reading the docs one can see that you are talking about parametrizations of ice shelf basal melt. Nonetheless, there are a lot of different physical processes involving melt on ice sheets (e.g. surface mass balance, basal mass blance, internal mass balance...). Perhaps being a little bit more specific would make things clearer for users.

### References

This is quite irrelevant, but I find the format of the references to be a little bit crude.

### Random

In the intro of the docs, it should read "at its heart" instead of "at it's heart".

I think it's very nice that you included so many different examples of increasing complexity. This really will help users to get started and understand the internals of the model. One little thing I'd fix there is the colorbar of the heatmaps. It's an annoying common issue in Julia, where you need to manually add an extra space in order to avoid an overlap between the title and the units.

### README

You have a very nice logo, why not add it to the README? 🙂

## Manuscript

Overall very nicely written and straight to the point. I really have almost no comments there, I think you managed to explain the field of ice sheet modelling to a wide audience, while providing enough details.

### Comparison to the field

You show examples of other models, but you don’t really mention what are the main differences between them. It is clear that WAVI.jl is more user-friendly, but perhaps it would be interesting to know a little bit more on why it is different or similar to other models.
