### 2023-07-14 #5584 @daniel-cheng
Hi @observingClouds, all, my checklist is up-to-date as well, but I am a bit delayed in completing it - I will be able to look over the manuscript next week though, and will should be able to evaluate the repository/code by the week after. I will be mindful of time though, and I'll be sure to update if I get things done sooner. Let me know if you need anything ahead of this schedule.
