### 2023-08-01 b9f15f2
Checks for approximate equality and equality
