+++ docs/src/examples/planar_1D_flow.md
@@ -13,7 +13,7 @@ First let's make sure we have all required packages installed.
 
 ```julia 
 using Pkg
-Pkg.add(url="https://github.com/RJArthern/WAVI.jl")
+Pkg.add(PackageSpec(url="https://github.com/RJArthern/WAVI.jl.git", rev = "main"))
 Pkg.add("Plots")
 using WAVI, Plots
 ```
