### 2023-09-26 #5584 @alextbradley
Hi @observingClouds (cc @PennyHow, @JordiBolibar, @daniel-cheng), thanks again for convening the review process and to the reviewers for their reports so far. I just wanted to give a progress update: the SparseArrays issue which prevented the reviewers from running the example is taking us a little longer to fix than we had expected. (We had only tested our code on Julia versions < 1.6.2, and this issue is arising only with newer versions of Julia, which all reviewers are using -- sorry about that!) We're confident that this should be fixed soon, and I'll let you know as soon as!
