+++ README.md
@@ -3,6 +3,8 @@
   ☃️🏔️❄️ WAVI.jl ❄️🏔️☃️
 </h1>
 
+![alt text](https://raw.githubusercontent.com/RJArthern/WAVI.jl/build-docs/docs/src/assets/logo.png)
+
 WAVI (Wavelet-based Adaptive-grid Vertically-integrated Ice-model) is a fast and friendly ice sheet model, written in Julia. WAVI documentation can be found [here](https://rjarthern.github.io/WAVI.jl/).
 
 ## Contents
