### 2023-10-16 #5584 @alextbradley
Hi @observingClouds @PennyHow @JordiBolibar @daniel-cheng, pleased to report that the bug causing the problems is now fixed (commit: https://github.com/RJArthern/WAVI.jl/commit/65a27dfbc8e5d0a939099bafb1c21abf2ff5c30c) 🎉 🥳. We've started to adjust the code and repo in response to the comments above -- @observingClouds, should we respond to those comments now, or wait for the rest of the evaluation of the reviewers?
