+++ paper.md
@@ -34,7 +34,7 @@ Physically, ice sheets do not stand alone, but are forced by other parts of the 
 
 `WAVI.jl` is the successor to a similar code, written in the proprietary programming language MATLAB, which was never publicly released. This previous code has been used extensively as research software [e.g. [@Arthern2017], [@Arthern2015]], as well as having participated in the most recent ice sheet model intercomparison exercise [@Cornford2020], which acts as a benchmark for ice-flow models. The new version, `WAVI.jl`, has also been verified independently against these benchmark experiments.
 
-There exists a wide variety of ice sheet models with varying levels of complexity. Examples include (but are certainly not limited to) the Ice Sheet System Model [@Larour2012], the Parallel Ice Sheet Model [@Bueler2009], BISICLES [@Cornford2013],  Elmer/Ice [@Gagliardini2013], and Úa [@Gudmundsson2019]. Every ice sheet model makes approximations in order to facilitate the numerical solution of the appropriate governing equations; since these equations have no analytic solutions, the _intercomparison_ between ice sheet models is of paramount importance when assessing the trustworthiness of models; `WAVI.jl` contributes to this community of ice sheet models. `WAVI.jl` is also, to our knowledge, the first ice sheet model written entirely in Julia and, alongside other accessible ice sheet models such as IcePack [@Shapero], helps to make ice sheet modelling more accessible.
+There exists a wide variety of ice sheet models with varying levels of complexity. Examples include (but are certainly not limited to) the Ice Sheet System Model [@Larour2012], the Parallel Ice Sheet Model [@Bueler2009], BISICLES [@Cornford2013],  Elmer/Ice [@Gagliardini2013], and Úa [@Gudmundsson2019]. Every ice sheet model makes approximations in order to facilitate the numerical solution of the appropriate governing equations; since these equations have no analytic solutions, the _intercomparison_ between ice sheet models is of paramount importance when assessing the trustworthiness of models; `WAVI.jl` contributes to this community of ice sheet models (a brief overview of the technical differences between such models can be found in [@Cornford2020]) `WAVI.jl` is also, to our knowledge, the first ice sheet model written entirely in Julia and, alongside other accessible ice sheet models such as IcePack [@Shapero], helps to make ice sheet modelling more accessible.
 
 # Acknowledgements
 
