+++ src/MeltRate/idealized_anthro_melt_rate.jl
@@ -1,0 +1,291 @@
+using Statistics
+
+struct IdealizedAnthroMeltRate{T <: Real, N <: Integer} <: AbstractMeltRate
+    bump_width :: T               #lengthscale of the quadratic bump in forcing
+    bump_time  :: T               #time of the center of the bump
+    bump_amplitude :: T           #amplitude of the bump
+    per_century_trend :: T        #per century trend in the forcing
+    trend_onset :: T              #time of onset of the trend
+    pc_max :: T                   #maximum pycnocline position (without a trend) 
+    pc_min :: T                   #minimum pycncoline position (without a trend)
+    M :: T                        #melt rate prefactor = γT .* (ρw * c / ρi /L)^2
+    λ1 :: T                       #liquidus slope 
+    λ2 :: T                       #liquidus intercept 
+    λ3 :: T                       #liquidus pressure coefficient
+    melt_partial_cell :: Bool     #Flag for melting applied to partial cells or not
+    random_seed :: N              #(integer) random seed for the stochastic process
+    rf_threshold :: T             #threshold for the autoregressive process to be capped at
+    r :: T                        #autocorrelation of AR1 process
+    Sl :: T                       #salinity of the lower layer 
+    Su :: T                       #salinity of the upper layer
+    Tl :: T                       #temperature of the upper layer
+    Tu :: T                       #temperature of the lower layer
+    pw :: T                       #width of the pycnocline
+end
+
+
+"""
+    IdealizedAnthroMeltRate(; ,kwargs)
+
+Construct a melt rate object for the idealized anthropogenic forcing experiments. This is a quadratic dependency of melting on depth with a variable prefactor, trend in forcing, and size of bump
+
+Keyword arguments 
+=================
+- bump_width :: T               #lengthscale of the quadratic bump in forcing
+- bump_time  :: T               #time of the center of the bump
+- bump_amplitude :: T           #amplitude of the bump
+- per_century_trend :: T        #per century trend in the forcing
+- trend_onset :: T              #time of onset of the trend
+- pc_max :: T                   #maximum pycnocline position (without a trend) 
+- pc_min :: T                   #minimum pycncoline position (without a trend)
+- M :: T                        #melt rate prefactor = γT .* (ρw * c / ρi /L)^2
+- λ1 :: T                       #liquidus slope 
+- λ2 :: T                       #liquidus intercept 
+- λ3 :: T                       #liquidus pressure coefficient
+- melt_partial_cell             #Flag for melting applied to partial cells or not
+- random_seed                   #random seed for the stochastic process
+- rf_threshold                  #threshold for the autoregressive process to be capped at
+- r                             #autocorrelation of AR1 process
+- Sl                            #salinity of the lower layer 
+- Su                            #salinity of the upper layer
+- Tl                            #temperature of the upper layer
+- Tu                            #temperature of the lower layer
+- pw                            #width of the pycnocline
+"""
+
+function IdealizedAnthroMeltRate(;
+                                bump_width = 1.0e-5,
+                                bump_time  = 0.0,
[…]
