+++ docs/make.jl
@@ -1,6 +1,7 @@
 push!(LOAD_PATH,"../src/")
 import Pkg; 
 Pkg.activate(joinpath(@__DIR__, ".."))
+Pkg.add(name="Documenter", version="0.27.25")
 Pkg.instantiate()
 
 using Documenter
@@ -107,12 +108,11 @@ format = Documenter.HTML(
 )
 
 
-makedocs(
+makedocs(bib,
     sitename = "WAVI.jl",
     format = format,
     pages = pages,
     modules = [WAVI],
-    plugins = [bib],
     doctest = false,
     strict = false,
     clean = false,
