+++ docs/make.jl
@@ -91,6 +91,7 @@ pages = [
     "API" => data_structure_pages,
    # "Simulation tips" => "simulation_tips.md",
     "MITgcm coupling" => "mitgcm_coupling.md",
+    "Contributors guide" => "contributing.md",
     "Contact us" => "contact.md",
     "References" => "references.md",
 ]
