### 2023-11-18 #5584 @alextbradley
Hi @PennyHow @observingClouds @JordiBolibar @daniel-cheng :) Here is our response to the outstanding comments so far!

@PennyHow

Issues highlighted pre-review:

1) Should this [bed = WAVI.mismip_plus_bed] line have some bracketed component?

    No, this generates the bed as a function, which can then be evaluated. In general, WAVI can accept numeric grids (i.e. where the bed is specified numerically at each grid point) or functions. The issue arising in the example was due to the bug highlighted in the next comment, which is now fixed.

2) I checked the other examples on your really nice documentation pages (https://rjarthern.github.io/WAVI.jl/). And I came across the same problem with this line in each of the examples. I'm not as familiar with Julia, so is this something I am doing wrong, something else I need in my set-up, or something that needs to be modified in your examples? Just FYI, I'm running Julia 1.9.0 on a Ubuntu 22.04 LTS machine.

    Thanks for highlighting this issue. As outlined in the previous comment (https://github.com/openjournals/joss-reviews/issues/5584#issuecomment-1764952630), this was caused by incompatibility with the sparse arrays package in newer versions of julia, which were untested and should have now been fixed.

3) I had to search hard for your documentation at https://rjarthern.github.io/WAVI.jl/ as I don't think it is included in your repo's README. Could you add this as a link to the repo, or a badge (if this is possible with your Documenter.jl set-up)? It has some great theoretical summaries, examples and walkthroughs.

    We've added a link in the readme and in the 'about' section of the repo, and sorry that it was hard to find!

4) With regard to the documentation, I have a request which I am happy to discuss with you the best way forward - Most packages use some form of autodocs to automatically generate documentation from the code (e.g. function arguments, outputs, and brief description). This would be highly beneficial to users to understand the capabilities and inputs to functions, troubleshoot problems, and adapt functions where needed. Could this be a possibility for you to include in your documentation?

    We have added these in the "WAVI setup" (now named "API") section of the docs for the key methods (Simulation, Model, TimesteppingParams, OutputParams, Grid, Params), using just the autogenerated internal docs. Does this look like what you had in mind? Happy to discuss further! Note that the way we have coded the InitialConditions and SolverParameters functions (using a slightly different structure, which doesn't affect user experience) precludes us from doing this. For those two, we have just described the constructors within the code above.

5) Great, easy-to-follow paper with a fantastic overview of current ice sheet modelling theory and available models. Just a couple of comments from me on this...

L8: I would say "lower the bar" is quite a colloqial term. Please change this to something more like "WAVI.jl is designed to make ice sheet modelling more accessible to beginners and low-level users..."
L25: What does the bracketed part of "On the long [O(1000s km)]..." mean"
L25: "lengthscales" >> "length scales"

    Thanks for these suggestions. In the updated paper, we have amended the 'lower the bar' statement as suggested (in L8 and in the final sentence), clarified that the bracketed part refers to "on the order of 1000s of kilometers", and fixed the "length scales" typo.

@JordiBolibar

- During installation, following your guidelines, I encountered the following warning...
└
There seems to be an issue with the package SparseArrays.jl, which as I will discuss in the next section, also hinders the use of your package.

I have tried running some of your examples (both the one in the README and the simplest one in the docs), but the code seems to crash when initializing the grid:

I would say this is by far the most important issue to address in the review. Probably it's just a minor technical issue with the package, but it would be important to have everything running smoothly before widely sharing the library with the community.

    Thanks for highlighting this issue. As outlined in the previous comment (https://github.com/openjournals/joss-reviews/issues/5584#issuecomment-1764952630), this was caused by incompatibility with the sparse arrays package in newer versions of julia, which were untested and should have now been fixed.

I'm not familiar with the format of the authors in JOSS, but to me in the manuscript there seems to be 3 authors, and in the repository and the code there seems to be mainly two authors. Is this normal? Probably I'm misunderstanding something here, so apologies if I got this wrong.

    It's true that most of the code has been written by myself and Rob, but other authors have gave active project direction and other forms of non-code contributions. We are happy to explicitly list these if needed (maybe @observingClouds can advise?)

I see this was also mentioned by @PennyHow, but I found it really hard to find the docs. Actually, I randomly came across them some weeks ago, but when I tried to look for the link in the repository during the review I just couldn't find it. I think a link and a mention of the docs should be clearly stated in the README (perhaps with a docs badge), and/or in the repository description.

    Sorry about this. As we outlined in our response to @PennyHow, we've now added a link to the docs in the 'about' section of the repo and in the first line of the readme.

The WAVI Setup section name is a little bit confusing to me. I feel it would be clearer to call it "API" or "WAVI data structures". Setup sounds more like configuration or installation to me.

    We have now renamed this section to API, thanks.

The image showing the data structures hierarchy is not displaying in the docs. It would be nice to fix this, since I'm sure it's really useful to users.

    Thanks for spotting this, this is now fixed.

The simulations section has some formatting issues, where some code is displayed as a title.

    Thanks for spotting this, this has now been corrected.

I find the term "melt rates" a bit vague. After reading the docs one can see that you are talking about parametrizations of ice shelf basal melt. Nonetheless, there are a lot of different physical processes involving melt on ice sheets (e.g. surface mass balance, basal mass blance, internal mass balance...). Perhaps being a little bit more specific would make things clearer for users.

    That's true, and sorry for the confusion. We've updated this to "basal melt rate parametrizations".

This is quite irrelevant, but I find the format of the references to be a little bit crude.

    We agree. Since we first formatted these, the DocumenterCitations package has added new formatting options (yay!) and we've adopted a numeric format now.

In the intro of the docs, it should read "at its heart" instead of "at it's heart".

    Thanks for spotting this typo, we've now fixed it.

I think it's very nice that you included so many different examples of increasing complexity. This really will help users to get started and understand the internals of the model. One little thing I'd fix there is the colorbar of the heatmaps. It's an annoying common issue in Julia, where you need to manually add an extra space in order to avoid an overlap between the title and the units.

    We're on this, and it will be fixed before the rest of the reviews come in.

You have a very nice logo, why not add it to the README? 🙂

    We have added the logo to the README :)

Overall very nicely written and straight to the point. I really have almost no comments there, I think you managed to explain the field of ice sheet modelling to a wide audience, while providing enough details.

    Thanks :)

You show examples of other models, but you don’t really mention what are the main differences between them. It is clear that WAVI.jl is more user-friendly, but perhaps it would be interesting to know a little bit more on why it is different or similar to other models.

    This is a bit tricky because it's hard to do this in a non-technical way and there are a number of different models to compare to. We've added a note that the interested reader can find details in Cornford et al. (2020) (https://doi.org/10.5194/tc-14-2283-2020), in which a brief comparison between most of the major ice sheet models can be found.

@daniel-cheng

General checks
The submission fulfills most checklist items, barring reproducibility of the results. I encountered similar issues as my fellow reviewers in attempting to run the code, which I was not able to resolve on my own. Functionality and performance of the code are not currently verified barring resolution of the code issues mentioned above.

    Thanks @daniel-cheng. As outlined above, this is now fixed and it should now run as expected!

There is a lack of clear contributing guidelines that should be added to the repository Read Me.

    We have added a contributors guide in the documentation.

There are tests, but not a clear way to run them in the publicly available documentation (as mentioned by the other reviewers at https://rjarthern.github.io/WAVI.jl/)

    We are working on this, and will add this to the documentation once the rest of the reviews come in.

The manuscript is clear and concise, and shows how the development of this Julia based ice sheet flow model enhances accessibility and scientific study by providing tools for the cryospheric community.

    Thanks :)
