### 2023-11-22 #5584 @observingClouds
Dear @alextbradley,
Thank you very much for addressing most of the reviewers comments and give us an update on the outstanding items, which I list for completeness and a better overview below:

- [ ] The colorbar of the heatmaps. It's an annoying common issue in Julia, where you need to manually add an extra space in order to avoid an overlap between the title and the units.
- [ ] Reproducibility of the results and general execution of the software
- [ ] There are tests, but not a clear way to run them in the publicly available documentation
- [x] Regarding the mis-match of authors on the paper and the repository: according to our [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) this is not an issue if these contributors provided "active project direction and other forms of non-code contributions" which @alextbradley confirmed for this submission.

Dear @PennyHow, @JordiBolibar, @daniel-cheng,
With the installation and usage issues potentially being fixed, I kindly ask you to test WAVE.jl's reproducability once more. Please also tick off all remaining checklist items if they were sufficiently addressed by @alextbradley's response.

Thank you all for your patience. I hope to make a final decision on this submission in two weeks time after getting your final evaluation.

Cheers,
Hauke
