### 2023-11-24 #5584 @PennyHow
Hi @alextbradley,

It works! Thanks for making those changes so that the examples work now. It's a very simple and straightforward set-up process. Just a couple of things from me now:

- Can you please make a PR and consolidate everything to the `main` branch. I see there are branches that contain the appropriate changes, but it's best to have them on the `main` branch as this is where users are first directed to.

- I also see that your Github pages build is held within a branch of WAVI.jl - it is best practise to limit branches to feature developments rather than entirely new projects/applications. I would advise moving the [gh-pages](https://github.com/RJArthern/WAVI.jl/tree/gh-pages) branch to a new repo. But this is not mandatory... theroetically you can use branches for anything!

- I've made some changes to your README as I found a typo and thought the Getting Help and Contributing sections looked a little empty. [Look at my PR](https://github.com/RJArthern/WAVI.jl/pull/75). Feel free to add more if needed. And then remember to merge the upstream branch to `main`

- With regards to documentation and the API section of the docs, I think this will suffice. It's not exactly what I had in mind, but I think it adequately describes how each aspect of WAVI.jl and its functionality operates.

Once these changes are made then I am happy to accept, @observingClouds
