### 2023-11-29 #5584 @JordiBolibar
Hi @alextbradley,

Thanks for addressing all the changes, great work! I have been able to successfully run the code, and the docs look really complete now.

My only remaining comments are the same ones as @PennyHow:

- I would make a PR from the `alextbradley-patch-1` branch into `main`, and I would make a new release of the model. I would encourage you to use a numbered version following the software release standards.
- I would also strongly encourage to move the GitHub pages to a new repo. Having it as a branch is a bit dangerous, and it makes maintenance and contributions from users more complicated.

Happy to accept the paper once these changes are made!
