+++ docs/src/examples/melt_parametrizations.md
@@ -16,7 +16,7 @@ Secondly, we demonstrate how to add a simple melt rate model to WAVI.jl.
 First let's make sure we have all required packages installed. As well as WAVI and Plots for plotting, we're going to use the Downloads package to pull some data from a Github repository.
 ```julia
 using Pkg
-pkg.add("https://github.com/RJArthern/WAVI.jl"), 
+Pkg.add(PackageSpec(url="https://github.com/RJArthern/WAVI.jl.git", rev = "main"))
 Pkg.add("Plots"), Pkg.add("Downloads")
 using WAVI, Plots, Downloads
 ```
