+++ docs/src/contributing.md
@@ -1,0 +1,55 @@
+# Contributors Guide
+
+Thank you for considering contributing to WAVI.jl! If you're interested in contributing to the development of WAVI we want your help no matter how big or small a contribution you make! 
+
+## Table of Contents
+- [How to Contribute](#how-to-contribute)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+  - [Code Contribution](#code-contribution)
+- [Development Setup](#development-setup)
+- [Pull Request Process](#pull-request-process)
+- [License](#license)
+
+## How to Contribute
+
+### Reporting Bugs
+
+If you encounter a bug, please help us fix it by following these steps:
+
+1. Ensure the bug is not already reported by checking the [issue tracker](https://github.com/RJArthern/WAVI.jl/issues).
+2. If the bug isn't reported, open a new issue. Clearly describe the issue, including steps to reproduce it.
+
+### Suggesting Enhancements
+
+If you have ideas for enhancements, new features, or improvements, we'd love to hear them! Follow these steps:
+
+1. Check the [issue tracker](https://github.com/RJArthern/WAVI.jl/issues) to see if your suggestion has been discussed.
+2. If not, open a new issue, providing a detailed description of your suggestion and the use case it addresses.
+
+### Code Contribution
+
+If you'd like to contribute code to the project:
+
+1. Fork the repository.
+2. Clone your fork: `git clone https://github.com/RJArthern/WAVI.jl`
+3. Create a new branch for your changes: `git checkout -b feature-branch`
+4. Make your changes and commit them with a clear message.
+5. Push your changes to your fork: `git push origin feature-branch`
+6. Open a pull request against the `main` branch of the main repository.
+
+## Pull Request Process
+
+Please ensure your pull request follows these guidelines:
+
+1. Adheres to the coding standards.
+2. Includes relevant tests for new functionality.
+3. Has a clear commit history and messages.
+4. References the relevant issue if applicable.
+
+Please don't hesistate to get in touch to discuss, or with any questions!
+
+## License
+
+By contributing to this project, you agree that your contributions will be licensed under the [LICENSE](https://github.com/RJArthern/WAVI.jl/blob/main/LICENSE) file of this repository.
+
