### 2024-01-10 05e06e3
fix docstring in OutputParams
