### 2024-01-10 adede4e
fix docstring in Grid
