### 2024-01-10 143d05b
fix bullets in Docs docs
