### 2024-01-10 e6ded9a
fix bullets in OutputParams docs
