### 2024-01-10 2683f64
fix bullets in Model docs
