+++ src/Models/Model.jl
@@ -22,14 +22,13 @@ Construct a WAVI.jl model object.
 
 Keyword arguments
 =================
-
-    - `grid`: (required) an instance of a `Grid` object, which defines the computational grid
-    - `bed_elevation`: (required) an array of size `grid.nx` x `grid.ny` which defines the bed elevation
-    - `params`: a `Params` object that defines physical parameters 
-    - `solver_params`: a `SolverParams` object that defines parameters relating to the numerical scheme
-    - `initial_conditions`: an `InitialConditions` object that (optionally) defines the initial ice thickness, temperature, viscosity, and damage
-    - `melt_rate`: a melt rate model, responsible for controlling/setting the basal melt rate
-    - `parallel_spec`: specification of parallel computation method.
+- `grid`: (required) an instance of a `Grid` object, which defines the computational grid
+- `bed_elevation`: (required) an array of size `grid.nx` x `grid.ny` which defines the bed elevation
+- `params`: a `Params` object that defines physical parameters 
+- `solver_params`: a `SolverParams` object that defines parameters relating to the numerical scheme
+- `initial_conditions`: an `InitialConditions` object that (optionally) defines the initial ice thickness, temperature, viscosity, and damage
+- `melt_rate`: a melt rate model, responsible for controlling/setting the basal melt rate
+- `parallel_spec`: specification of parallel computation method.
 
 """
 function Model(;
