### 2024-01-10 090dcee
fix Simulation docs bullet points
