+++ src/Params.jl
@@ -32,29 +32,29 @@ Construct a WAVI.jl parameters object for holding physical parameters.
 
 Keyword arguments
 =================
-- 'dt': model timestep (NB: simulation timestep set in timestepping params, this value is updated when model embedded to the value specified in timestepping_params when passed to simulation)
-- 'g': gravitational acceleration (m^2 / s)
-- 'density_ice': ice density (kg / m^3)
-- 'density_ocean' ocean water density (kg / m^3)
-- 'gas_const': gas constant in glen b calculation
-- 'sec_per_year': seconds per year (s)
-- 'default_thickness': thickness value reverted to if no initial thickness passed (m)
-- 'default_viscosity': viscosity value reverted to if no initial thickness passed (Pa s)
-- 'default_temperature': temperature value reverted to if no initial thickness passed (K)
-- 'default_damage': damage value reverted to if no initial thickness passed (dimensionless)
-- 'accumulation_rate': uniform accumulation_rate (m/yr)
-- 'glen_a_activation_energy': activation energy in glen b calculation
-- 'glen_a_ref': array of glen a reference values used in glen b calculation
-- 'glen_temperature_ref': reference temperature using in glen b calculation
-- 'glen_n': exponent in glen b calculation
-- 'glen_reg_strain_rate': strain rate regularization value
-- 'weertman_c': basal sliding field of coefficients
-- 'weertman_m': sliding law exponent
-- 'weertman_reg_speed': regularization speed, used to prevent bed speed going to zero
-- 'sea_level_wrt_geoid': reference sea level
-- 'minimum_thickness': minimum ice thickness on model domain
-- 'evolveShelves': flag for turning on and off the evolution of the shelves in the forward run_simulation
-- 'smallHAF': small value of HAF used within update_thickness when not evolving shelves
+- `dt`: model timestep (NB: simulation timestep set in timestepping params, this value is updated when model embedded to the value specified in timestepping_params when passed to simulation)
+- `g`: gravitational acceleration (m^2 / s)
+- `density_ice`: ice density (kg / m^3)
+- `density_ocean`: ocean water density (kg / m^3)
+- `gas_const`: gas constant in glen b calculation
+- `sec_per_year`: seconds per year (s)
+- `default_thickness`: thickness value reverted to if no initial thickness passed (m)
+- `default_viscosity`: viscosity value reverted to if no initial thickness passed (Pa s)
+- `default_temperature`: temperature value reverted to if no initial thickness passed (K)
+- `default_damage`: damage value reverted to if no initial thickness passed (dimensionless)
+- `accumulation_rate`: uniform accumulation_rate (m/yr)
+- `glen_a_activation_energy`: activation energy in glen b calculation
+- `glen_a_ref`: array of glen a reference values used in glen b calculation
+- `glen_temperature_ref`: reference temperature using in glen b calculation
+- `glen_n`: exponent in glen b calculation
+- `glen_reg_strain_rate`: strain rate regularization value
+- `weertman_c`: basal sliding field of coefficients
+- `weertman_m`: sliding law exponent
+- `weertman_reg_speed`: regularization speed, used to prevent bed speed going to zero
+- `sea_level_wrt_geoid`: reference sea level
+- `minimum_thickness`: minimum ice thickness on model domain
+- `evolveShelves`: flag for turning on and off the evolution of the shelves in the forward run_simulation
+- `smallHAF`: small value of HAF used within update_thickness when not evolving shelves
 """
 function Params(; g = 9.81, 
                   density_ice = 918.0,
