### 2024-01-10 aa4c1b0
Update melt_rate_models.md
