+++ .github/workflows/Documentation.yml
@@ -3,7 +3,7 @@ name: Documentation
 on:
   push:
     branches:
-      - build-docs # update to match your development branch (master, main, dev, trunk, ...)
+      - main # update to match your development branch (master, main, dev, trunk, ...)
     tags: '*'
   pull_request:
 
