+++ .github/workflows/Testing.yml
@@ -1,0 +1,27 @@
+name: Testing
+
+on:
+  push:
+    branches:
+      - docs-reconcile # update to match your development branch (master, main, dev, trunk, ...)
+    tags: '*'
+  pull_request:
+
+jobs:
+  build:
+    permissions:
+      contents: write
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: julia-actions/setup-julia@v1
+        with:
+          version:
+          - '1.6'
+          - '1' #'1' will automatically expand to the latest stable 1.x release of Julia.
+      - name: Install dependencies
+        run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+      - name: Build and deploy
+        env:
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }} # If authenticating with SSH deploy key
+        run: julia --project=test/ test/runtests_.jl
