+++ .github/workflows/Testing.yml
@@ -16,9 +16,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
         with:
-          version:
-          - '1.6'
-          - '1' #'1' will automatically expand to the latest stable 1.x release of Julia.
+          version: '1.6'
       - name: Install dependencies
         run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
       - name: Build and deploy
