### 2024-01-11 43c8f4a
add structure only tests for CI
