+++ .github/workflows/Documentation.yml
@@ -3,7 +3,8 @@ name: Documentation
 on:
   push:
     branches:
-      - [docs-reconcile, main] # update to match your development branch (master, main, dev, trunk, ...)
+      - docs-reconcile
+      - main # update to match your development branch (master, main, dev, trunk, ...)
     tags: '*'
   pull_request:
 
