+++ README.md
@@ -1,4 +1,6 @@
-![example branch parameter](https://github.com/RJArthern/WAVI.jl/actions/workflows/Testing.yml/badge.svg?docs-reconcile)
+![Testing](https://github.com/RJArthern/WAVI.jl/actions/workflows/Testing.yml/badge.svg?docs-reconcile)
+
+![Documentation](https://github.com/RJArthern/WAVI.jl/actions/workflows/Documentation.yml/badge.svg?docs-reconcile)
 
 <!-- Title -->
 <h1 align="center">
