### 2024-01-14 e79deeb
Update WAIS.md
fix typo
