### 2024-01-14 6b3cfc2
fix colorbars in bumpy bed
