+++ docs/src/examples/bumpy_bed.md
@@ -10,7 +10,7 @@ First let's make sure we have all required packages installed.
 ```julia
 using Pkg
 Pkg.add(PackageSpec(url="https://github.com/RJArthern/WAVI.jl.git", rev = "main"))
-Pkg.add(`Plots`)
+Pkg.add("Plots")
 using WAVI, Plots
 ```
 
@@ -48,7 +48,7 @@ plot!(size = (800,800))
 ```
 
 ```@raw html
-<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/build-docs/docs/src/assets/example-plots//bumpy//bed.png" alt="" title="" width="600" height="600" /></center>
+<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/docs-reconcile/docs/src/assets/example-plots//bumpy//bed.png" alt="" title="" width="600" height="600" /></center>
 ```
 
 ## Model Instantiation and Initial Conditions
@@ -81,7 +81,7 @@ plot!(size = (800,600))
 ```
 
 ```@raw html
-<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/build-docs/docs/src/assets/example-plots//bumpy//velocity_L80.png" alt="" title="" width="600" height="600" /></center>
+<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/docs-reconcile/docs/src/assets/example-plots//bumpy//velocity_L80.png" alt="" title="" width="600" height="600" /></center>
 ```
 
 ## Different lengthscales
@@ -126,7 +126,7 @@ display(p)
 plot!(size = (1000,550))
 ```
 ```@raw html
-<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/build-docs/docs/src/assets/example-plots//bumpy//velocity_diffL.png" alt="" title="" width="600" height="600" /></center>
+<center><img src="https://raw.githubusercontent.com/RJArthern/WAVI.jl/docs-reconcile/docs/src/assets/example-plots//bumpy//velocity_diffL.png" alt="" title="" width="600" height="600" /></center>
 ```
 
 As expected, when the bumps have a smaller aspect ratio (smaller `L`) the flow speed is smaller.
