### 2024-01-14 81fd4f7
update colorbar locations in mismip plus bed
