### 2024-01-16 9eb8e95
fix mask error in melt rate specification
