+++ tmp/driver.jl
@@ -1,0 +1,184 @@
+using WAVI 
+function driver()
+
+
+#
+#Grid and boundary conditions
+#
+nx = 398
+ny = 250
+nσ = 12
+x0 = -1789000.0
+y0 = -833000.0
+dx = 2000.0
+dy = 2000.0
+
+h_mask=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_h_mask_clip_BedmachineV3_FULL_stripe_fix.bin",h_mask)
+h_mask.=ntoh.(h_mask)
+
+#
+# boundary conditions
+#
+u_iszero=Array{Float64}(undef,nx+1,ny);
+read!("Inverse_2km_uiszero_clip_BedmachineV3_FULL_stripe_fix.bin",u_iszero)
+u_iszero.=ntoh.(u_iszero)
+
+v_iszero=Array{Float64}(undef,nx,ny+1);
+read!("Inverse_2km_viszero_clip_BedmachineV3_FULL_stripe_fix.bin",v_iszero)
+v_iszero.=ntoh.(v_iszero)
+
+sigma_grid=Array{Float64}(undef,nσ);
+read!("Inverse_2km_sigma_grid_BedmachineV3_FULL_stripe_fix.bin",sigma_grid)
+sigma_grid.=ntoh.(sigma_grid)
+
+grid = Grid(nx = nx,
+            ny = ny,
+            nσ = nσ,
+            x0 = x0,
+            y0 = y0,
+            dx = dx,
+            dy = dy,
+            h_mask = h_mask,
+            u_iszero = u_iszero,
+            v_iszero = v_iszero,
+            σ = sigma_grid)
+
+#
+#input files
+#
+bed=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_bed_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",bed)
+bed.=ntoh.(bed)
+
+h=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_thickness_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",h)
+h.=ntoh.(h)
+
+viscosity=Array{Float64}(undef,nx,ny,nσ);
+read!("Inverse_2km_viscosity3D_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",viscosity)
+viscosity.=ntoh.(viscosity)
+
+temp=Array{Float64}(undef,nx,ny,nσ);
+read!("Inverse_2km_3Dtemp_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",temp)
+temp.=ntoh.(temp)
+
+damage=Array{Float64}(undef,nx,ny,nσ);
+read!("Inverse_2km_damage3D_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",damage)
+damage.=ntoh.(damage)
+
+weertman_c=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_WeertmanC_clip_adjusted_noNan_BedmachineV3_FULL_stripe_fix.bin",weertman_c)
+weertman_c.=ntoh.(weertman_c)
+
+accumulation_rate=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_accumulation_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",accumulation_rate)
+accumulation_rate.=ntoh.(accumulation_rate)
+
+dhdt=Array{Float64}(undef,nx,ny);
+read!("Inverse_2km_dhdt_clip_noNan_BedmachineV3_FULL_stripe_fix.bin",dhdt)
+dhdt.=ntoh.(dhdt)
+
+initial_conditions = InitialConditions(initial_thickness = h,
+                                        initial_viscosity = viscosity,
[…]
