### 2024-02-08 669dc49
add thickness substepping capability
