+++ src/Models/update_state.jl
@@ -41,6 +41,25 @@ function update_state!(model)
     return nothing
 end
 
+"""
+update_state_novelocity!(model::AbstractModel, clock)
+
+Update the model to the current time dependent situation, without updating the ice velocities
+"""
+function update_state_novelocity!(model, clock)
+    update_surface_elevation!(model)
+    update_geometry_on_uv_grids!(model)
+    update_height_above_floatation!(model)
+    update_grounded_fraction_on_huv_grids!(model)
+    update_accumulation_rate!(model)
+    update_basal_melt!(model, clock)
+    update_weertman_c!(model)
+    update_dsdh!(model)
+    update_velocities_on_h_grid!(model)
+    update_dhdt!(model)
+    update_model_wavelets!(model)
+    return nothing
+end
 
 
 """
