### 2024-02-09 #5584 @alextbradley
Hi folks, sorry for the delay in responding to this. We've now addressed all of the above comments, which were very helpful in improving the functionality of the software and streamlining the user experience. I'll respond to them in detail here:

> The colorbar of the heatmaps. It's an annoying common issue in Julia, where you need to manually add an extra space in order to avoid an overlap between the title and the units.

Thanks for spotting this — we’ve now fixed this issue in the examples.

> Reproducibility of the results and general execution of the software

I think that Jordi and Penny have now confirmed that they are able to run the software, and we are able to run on different architectures, so hopefully this comment is now addressed.

> There are tests, but not a clear way to run them in the publicly available documentation

We have added a section in the documentation describing the tests and how to run them using Julia’s unit testing architecture. We’ve also introduce continuous integration (as suggested by Jordi) to automatically test the software.

> Regarding the mis-match of authors on the paper and the repository: according to our guidelines this is not an issue if these contributors provided "active project direction and other forms of non-code contributions" which @alextbradley confirmed for this submission. Please note however that the author list of the final archive (e.g. zenodo release) has to agree with the list of authors on the JOSS paper.

We will issue a zenodo release and a new GitHub release when the paper is accepted and this will have the same list of authors as the JOSS paper.

> Can you please make a PR and consolidate everything to the main branch. I see there are branches that contain the appropriate changes, but it's best to have them on the main branch as this is where users are first directed to.

We have now consolidated the changes to the main branch. (The reason we had been doing this is that we don’t allow direct commits to main without review, and so we were developing the code in a branch with the intention of merging back, which we’ve now done.)

> I also see that your Github pages build is held within a branch of WAVI.jl - it is best practise to limit branches to feature developments rather than entirely new projects/applications. I would advise moving the gh-pages branch to a new repo. But this is not mandatory... theroetically you can use branches for anything!

Yes, thanks for flagging this! The build of the documentation is now in the main branch, which we hope this is an acceptable solution. It’s a bit less hassle to keep them in the same place, but if there’s a good reason for moving to a new repo rather than a main branch we’re more than happy to engage! It should be noted that the gh-pages branch is autogenerated by the GitHub action for Documenter.jl (the package we use for building the docs) which builds the documentation based on what is in main branch and then deployed.

> I've made some changes to your README as I found a typo and thought the Getting Help and Contributing sections looked a little empty. Look at my PR. Feel free to add more if needed. And then remember to merge the upstream branch to main

Thank-you for these changes! We’ve merged your PR.

> With regards to documentation and the API section of the docs, I think this will suffice. It's not exactly what I had in mind, but I think it adequately describes how each aspect of WAVI.jl and its functionality operates.

Okay, great!

> I would make a PR from the alextbradley-patch-1 branch into main, and I would make a new release of the model. I would encourage you to use a numbered version following the software release standards.

Sorry, we had opened this as a PR, but had yet to merge it. We’ve now done this!

> I would also strongly encourage to move the GitHub pages to a new repo. Having it as a branch is a bit dangerous, and it makes maintenance and contributions from users more complicated.

See above comment in response to Penny. We’ve now moved to main, which cannot be committed to without review.

> I have tried running the tests from package mode, but I was not able to do so. Have you set up a test suite? I also see that there are no CI tests set up in the GitHub repository. This is not mandatory, but for a community model it would be very practical to make sure that any incoming PR doesn't break the codebase.

We have now added details of the testing to documentation, and added CI via a GitHub action. We’ve also added a badge on the readme showing the status of the testing action.
