### 2024-02-23 #5584 @alextbradley
Hi @observingClouds -- a comments of questions with the proof:
- the author list is not correct, and should read "Alexander T. Bradley, Robert Arthern, David T. Bett, C. Rosie Williams, James Byrne" (all have the same affiliation" "British Antarctic Survey, Cambridge, UK"). It renders correctly in our github action which builds the pdf, so I'm not sure what has happened there.
- two small typos (our fault): there should be a full stop after " (a brief overview of the technical
68 differences between such models can be found in (Cornford et al., 2020)) " and there should be a space in line 38, i.e. "which fringe this region (Pritchard et al., 2012)."
