### 2024-02-23 #5584 @observingClouds
Hi @alextbradley,
to you comments:
- all authors have been listed under one `name` entry. As a consequence the authors in the rendered paper are not separated by a comma. Please create an entry for each author as suggested in my PR.
- please edit the paper.md accordingly. It seems like they overlap with my suggestions in https://github.com/RJArthern/WAVI.jl/pull/78 :+1:
