### 2024-03-05 #5584 @observingClouds
Hi @alextbradley,
I need to ask you to do a few more things before we can publish the paper:
- [ ] add the missing DOI in the references of the paper (see [reference check](https://github.com/openjournals/joss-reviews/issues/5584#issuecomment-1979300975))
- [ ] please adjust the title of the archive to the title of the paper (`RJArthern/WAVI.jl: zenodo release` --> `WAVI.jl: Ice Sheet Modelling in Julia`)
- [ ] I also encourage you to add your ORCiD and those of the co-authors to the manuscript and the zenodo release
