+++ paper.bib
@@ -213,7 +213,8 @@
   number={4},
   pages={1299--1318},
   year={2013},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/gmd-6-1299-2013}
 }
 
 @misc{Gudmundsson2019,
