### 2024-03-08 #5584 @alextbradley
Hi @observingClouds, these should now be fixed! I'll just generate the pdf with the editorial bot to check the DOI and ORCiDs in the paper are correct. Zenodo has been updated to have correct title and ORCiDs (https://zenodo.org/records/10723504)
