+++ paper.bib
@@ -221,7 +221,8 @@
   title = {Úa: A large-scale ice-flow model},
   howpublished = {\url{https://github.com/GHilmarG/UaSource}},
   author = {Gudmundsson, H G},
-  year = {2019}
+  year = {2019},
+  doi={10.5281/zenodo.3706624}
 }
 
 @article{Bamber2018,
@@ -249,74 +250,3 @@
 }
 
 
-@article{Pearson:2017,
-	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-	Adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-	Archiveprefix = {arXiv},
-	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-	Eprint = {1703.04627},
-	Journal = {ArXiv e-prints},
-	Keywords = {Astrophysics - Astrophysics of Galaxies},
-	Month = mar,
-	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-	Year = 2017}
-
-@book{Binney:2008,
-	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-	Adsurl = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-	Author = {{Binney}, J. and {Tremaine}, S.},
-	Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
-	Publisher = {Princeton University Press},
-	Title = {{Galactic Dynamics: Second Edition}},
-	Year = 2008}
-
-@ARTICLE{gaia,
-   author = {{Gaia Collaboration} and {Prusti}, T. and {de Bruijne}, J.~H.~J. and
-	{Brown}, A.~G.~A. and {Vallenari}, A. and {Babusiaux}, C. and
-	{Bailer-Jones}, C.~A.~L. and {Bastian}, U. and {Biermann}, M. and
-	{Evans}, D.~W. and et al.},
-    title = "{The Gaia mission}",
-  journal = {\aap},
-archivePrefix = "arXiv",
-   eprint = {1609.04153},
- primaryClass = "astro-ph.IM",
- keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
-     year = 2016,
-    month = nov,
-   volume = 595,
-      eid = {A1},
-    pages = {A1},
-      doi = {10.1051/0004-6361/201629272},
-   adsurl = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{astropy,
-   author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. and
-	{Greenfield}, P. and {Droettboom}, M. and {Bray}, E. and {Aldcroft}, T. and
-	{Davis}, M. and {Ginsburg}, A. and {Price-Whelan}, A.~M. and
-	{Kerzendorf}, W.~E. and {Conley}, A. and {Crighton}, N. and
-	{Barbary}, K. and {Muna}, D. and {Ferguson}, H. and {Grollier}, F. and
-	{Parikh}, M.~M. and {Nair}, P.~H. and {Unther}, H.~M. and {Deil}, C. and
-	{Woillez}, J. and {Conseil}, S. and {Kramer}, R. and {Turner}, J.~E.~H. and
[…]
