### 2024-03-12 #5584 @kthyng
Paper:
* last paragraph first page: space before "Pritchard" reference
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
