+++ paper.md
@@ -38,7 +38,7 @@ Ice sheet models allow us to simulate the behaviour and evolution of ice sheets,
 
 On the long (100s of kilometers) length scales that are relevant to ice sheets, ice behaves approximately as a highly viscous fluid with a shear-thinning rheology (meaning that as the ice deforms, it becomes thinner and flows more easily). `WAVI.jl` (Wavelet-based Adaptive-grid Vertically-integrated Ice-sheet-model) is a Julia package for the numerical solution of an accurate approximation to the Stokes equations, which describe conservation of mass and momentum in such a fluid. This approximation, which is appropriate for fluid flows with a high aspect ratio (as is the case for the vast majority of ice sheets), treats longitudinal and lateral stresses as depth-independent, but accounts for vertical velocity gradients in the nonlinear viscosity and in the treatment of basal stress [@Goldberg2011].
 
-Physically, ice sheets do not stand alone, but are forced by other parts of the climate system. For example, the rapid changes that have occurred in the West Antarctic Ice Sheet in the previous decades are understood to have been driven by an increase in oceanic heat content reaching the floating ice shelves which fringe this region[@Pritchard2012]. For basal melting in particular, `WAVI.jl` includes a broad range of community melt rate parametrizations [@AsayDavis2017], as well as a developmental coupling with the ocean general circulation model MITgcm [@Marshall1997]. More generally, `WAVI.jl` leverages Julia's multiple dispatch paradigm to create a simple, user-friendly interface for embedding models of other physical processes, such as accumulation, ice damage, ice shelf calving, and solid earth effects, into `WAVI.jl`.
+Physically, ice sheets do not stand alone, but are forced by other parts of the climate system. For example, the rapid changes that have occurred in the West Antarctic Ice Sheet in the previous decades are understood to have been driven by an increase in oceanic heat content reaching the floating ice shelves which fringe this region [@Pritchard2012]. For basal melting in particular, `WAVI.jl` includes a broad range of community melt rate parametrizations [@AsayDavis2017], as well as a developmental coupling with the ocean general circulation model MITgcm [@Marshall1997]. More generally, `WAVI.jl` leverages Julia's multiple dispatch paradigm to create a simple, user-friendly interface for embedding models of other physical processes, such as accumulation, ice damage, ice shelf calving, and solid earth effects, into `WAVI.jl`.
 
 `WAVI.jl` is designed to be usable by anyone interested in ice sheet modelling, from students with no programming experience to expert researchers in the field, and everyone in between. To facilitate detailed research, including simulations at high spatial and temporal resolution, `WAVI.jl` employs a number of tools to improve computational speed, including multithreading capabilities, an adaptive numerical grid and a wavelet-based preconditioner [@Arthern2017]. To facilitate accessibility, `WAVI.jl` includes a simple, user friendly API, which is aided by Julia's convenient syntax. In addition, the GitHub repository in which the code is stored includes a number of well-documented examples, which demonstrate the software's capabilities in a wide variety of situations.
 
