+++ paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{Favier2014,
-  title={Retreat of Pine Island Glacier controlled by marine ice-sheet instability},
+  title={Retreat of {P}ine {I}sland {G}lacier controlled by marine ice-sheet instability},
   author={Favier, Lionel and Durand, Gael and Cornford, Stephen L and Gudmundsson, G Hilmar and Gagliardini, Olivier and Gillet-Chaulet, Fabien and Zwinger, Thomas and Payne, AJ and Le Brocq, Anne M},
   journal={Nature Climate Change},
   volume={4},
