+++ paper.bib
@@ -25,7 +25,7 @@
 
 
 @article{DeConto2021,
-  title={The Paris Climate Agreement and future sea-level rise from Antarctica},
+  title={The {P}aris {C}limate {A}greement and future sea-level rise from {A}ntarctica},
   author={DeConto, Robert M and Pollard, David and Alley, Richard B and Velicogna, Isabella and Gasson, Edward and Gomez, Natalya and Sadai, Shaina and Condron, Alan and Gilford, Daniel M and Ashe, Erica L and others},
   journal={Nature},
   volume={593},
@@ -37,7 +37,7 @@
 }
 
 @article{DeRydt2021,
-  title={Drivers of Pine Island Glacier speed-up between 1996 and 2016},
+  title={Drivers of {P}ine {I}sland {G}lacier speed-up between 1996 and 2016},
   author={De Rydt, Jan and Reese, Ronja and Paolo, Fernando S and Gudmundsson, G Hilmar},
   journal={The Cryosphere},
   volume={15},
@@ -55,7 +55,7 @@
                   David Lilien and
                   Andrew Hoffman},
   title        = {{icepack/icepack: icepack: glacier flow modeling
-                   with the finite element method in Python}},
+                   with the finite element method in {P}ython}},
   month        = dec,
   year         = 2020,
   publisher    = {Zenodo},
@@ -65,7 +65,7 @@
 }
 
 @article{Joughin2021,
-  title={Ice-shelf retreat drives recent Pine Island Glacier speedup},
+  title={Ice-shelf retreat drives recent {P}ine {I}sland {G}lacier speedup},
   author={Joughin, Ian and Shapero, Daniel and Smith, Ben and Dutrieux, Pierre and Barham, Mark},
   journal={Science Advances},
   volume={7},
@@ -88,7 +88,7 @@
 }
 
 @article{Arthern2015,
-  title={Flow speed within the Antarctic ice sheet and its controls inferred from satellite observations},
+  title={Flow speed within the {A}ntarctic ice sheet and its controls inferred from satellite observations},
   author={Arthern, Robert J and Hindmarsh, Richard CA and Williams, C Rosie},
   journal={Journal of Geophysical Research: Earth Surface},
   volume={120},
@@ -100,7 +100,7 @@
 }
 
 @article{Arthern2017,
-  title={The sensitivity of West Antarctica to the submarine melting feedback},
+  title={The sensitivity of {W}est {A}ntarctica to the submarine melting feedback},
   author={Arthern, Robert J and Williams, C Rosie},
   journal={Geophysical Research Letters},
   volume={44},
@@ -124,7 +124,7 @@
 }
 
 @article{AsayDavis2017,
-  title={Developments in simulating and parameterizing interactions between the Southern Ocean and the Antarctic ice sheet},
+  title={Developments in simulating and parameterizing interactions between the {S}outhern {O}cean and the {A}ntarctic ice sheet},
   author={Asay-Davis, Xylar S and Jourdain, Nicolas C and Nakayama, Yoshihiro},
   journal={Current Climate Change Reports},
   volume={3},
@@ -160,7 +160,7 @@
 }
 
 @article{Cornford2020,
-  title={Results of the third Marine Ice Sheet Model Intercomparison Project (MISMIP+)},
+  title={Results of the third {M}arine {I}ce {S}heet {M}odel {I}ntercomparison {P}roject (MISMIP+)},
[…]
