+++ paper.bib
@@ -54,7 +54,7 @@
                   David A. Ham and
                   David Lilien and
                   Andrew Hoffman},
-  title        = {{icepack/icepack: icepack: glacier flow modeling
+  title        = {{icepack: glacier flow modeling
                    with the finite element method in {P}ython}},
   month        = dec,
   year         = 2020,
